[FE] 파일 위치 조정
