+++ src/components/common/CreateCardModal.tsx
@@ -3,44 +3,63 @@ import { useNavigate } from 'react-router-dom';
 import {
   Modal,
   Box,
+  Checkbox,
   TextField,
   Button,
   Select,
   MenuItem,
   InputLabel,
   FormControl,
+  FormControlLabel,
   Grid,
   SelectChangeEvent
 } from '@mui/material';
 import { useTheme } from '@mui/material/styles';
-import { getAllProductCategory, getAllArea } from '../api/axios.custom.ts';
+import {
+  getAllProductCategory,
+  getAllArea,
+  createCard
+} from '../../api/axios.custom.ts';
 import {
   AllCategory,
   AllSubCategory,
   AllSido,
-  AllSigg
-} from '../types/types.ts';
+  AllSigg,
+  ProductSearchRequest
+} from '../../types/types.ts';
 
-interface SearchModalProps {
+interface CreateCardModalProps {
   open: boolean;
   handleClose: () => void;
+  searchOption: ProductSearchRequest;
 }
 
-const SearchModal: React.FC<SearchModalProps> = ({ open, handleClose }) => {
-  const [searchTerm, setSearchTerm] = useState('');
-  const [minPrice, setMinPrice] = useState('');
-  const [maxPrice, setMaxPrice] = useState('');
+const CreateCardModal: React.FC<CreateCardModalProps> = ({
+  open,
+  handleClose,
+  searchOption
+}) => {
+  const [newCardTitle, setNewCardTitle] = useState('');
+  const [searchTerm, setSearchTerm] = useState(searchOption.keyword);
+  const [minPrice, setMinPrice] = useState(
+    searchOption.minPrice ? searchOption.minPrice.toString() : ''
+  );
+  const [maxPrice, setMaxPrice] = useState(
+    searchOption.maxPrice ? searchOption.maxPrice.toString() : ''
+  );
   const [category, setCategory] = useState<AllCategory[]>([]);
   const [subCategory, setSubCategory] = useState<AllSubCategory[]>([]);
-  const [selectedSubCategory, setSelectedSubCategory] = useState<string>('');
-  const [selectedCategory, setSelectedCategory] = useState<string>('');
   const [selectedSuperCategory, setSelectedSuperCategory] =
     useState<string>('');
+  const [selectedSubCategory, setSelectedSubCategory] = useState<string>('');
+  const [selectedCategory, setSelectedCategory] = useState<string>('');
   const [area, setArea] = useState<AllSido[]>([]);
   const [selectedSido, setSelectedSido] = useState<AllSido>();
   const [selectedSigg, setSelectedSigg] = useState<AllSigg>();
   const [selectedEmd, setSelectedEmd] = useState<number>();
+  const [isPublic, setIsPublic] = useState(true);
   const navigate = useNavigate();
+
   useEffect(() => {
     const fetchCategoriesAndAreas = async () => {
       try {
@@ -49,11 +68,18 @@ const SearchModal: React.FC<SearchModalProps> = ({ open, handleClose }) => {
         const response2 = await getAllArea();
         setArea(response2.data.areas[0].sido);
       } catch (error) {
-        console.log('카테고리 데이터를 불러오는데 실패했습니다.', error);
+        console.error('카테고리 데이터를 불러오는데 실패했습니다.', error);
       }
     };
     fetchCategoriesAndAreas();
   }, []);
+
+  const handleNewCardTitleChange = (
+    event: React.ChangeEvent<HTMLInputElement>
+  ) => {
+    setNewCardTitle(event.target.value);
+  };
+
   const handleSearchTermChange = (
     event: React.ChangeEvent<HTMLInputElement>
   ) => {
@@ -70,21 +96,13 @@ const SearchModal: React.FC<SearchModalProps> = ({ open, handleClose }) => {
 
   const handleCategoryChange = (event: SelectChangeEvent<string>) => {
     const categoryId = event.target.value as string;
-
-    if (categoryId === 'default') {
-      setSelectedSuperCategory('');
-      setSelectedCategory('');
-      setSelectedSubCategory('');
-      setSubCategory([]);
-    } else {
-      setSelectedSuperCategory(categoryId);
-      setSelectedCategory(categoryId);
-      const foundCategory = category.find(
-        (c) => c.categoryId === Number(categoryId)
-      );
-      setSubCategory(foundCategory?.subCategory || []);
-      setSelectedSubCategory('default');
-    }
+    setSelectedSuperCategory(categoryId);
+    setSelectedCategory(categoryId);
+    const foundCategory = category.find(
+      (c) => c.categoryId === Number(categoryId)
+    );
+    setSubCategory(foundCategory?.subCategory || []);
+    setSelectedSubCategory('default');
   };
 
   const handleSubCategoryChange = (event: SelectChangeEvent<string>) => {
@@ -93,22 +111,15 @@ const SearchModal: React.FC<SearchModalProps> = ({ open, handleClose }) => {
     if (subCategoryId !== 'default') {
       setSelectedCategory(subCategoryId);
     } else {
-      setSelectedSubCategory('');
+      setSelectedCategory(selectedSuperCategory);
     }
   };
 
   const handleSidoSelect = (event: SelectChangeEvent) => {
     const sidoId = event.target.value;
-    if (sidoId === 'default') {
-      setSelectedSido(undefined);
-      setSelectedSigg(undefined);
-      setSelectedEmd(undefined);
-    } else {
-      const selected = area.find((sido) => sido.id === Number(sidoId));
-      setSelectedSido(selected);
-      setSelectedSigg(undefined);
-      setSelectedEmd(undefined);
-    }
+    const selected = area.find((sido) => sido.id === Number(sidoId));
+    setSelectedSido(selected);
+    setSelectedEmd(undefined);
   };
 
   const handleSiggSelect = (event: SelectChangeEvent) => {
@@ -124,15 +135,38 @@ const SearchModal: React.FC<SearchModalProps> = ({ open, handleClose }) => {
     setSelectedEmd(Number(emdId));
   };
 
-  const handleSearch = () => {
-    const searchParams = new URLSearchParams(); // 검색 조건을 URL 쿼리 파라미터로 변환
-    searchParams.append('keyword', encodeURIComponent(searchTerm));
-    if (minPrice) searchParams.append('minPrice', minPrice);
-    if (maxPrice) searchParams.append('maxPrice', maxPrice);
-    if (selectedCategory) searchParams.append('categoryId', selectedCategory);
-    if (selectedEmd) searchParams.append('areaId', selectedEmd.toString());
-    navigate(`/searchresult?${searchParams.toString()}`);
-    handleClose();
+  const handlePublicChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+    setIsPublic(event.target.checked);
+  };
+
+  const handleCreate = async () => {
+    const formData = new FormData();
+
+    // Append only if values are not undefined
+    if (selectedCategory) {
+      formData.append('categoryId', selectedCategory);
+    }
+    if (selectedEmd !== undefined) {
+      formData.append('areaId', selectedEmd.toString());
+    }
+    formData.append('title', newCardTitle);
+    formData.append('keyword', searchTerm);
+    if (minPrice) {
+      formData.append('minPrice', minPrice);
+    }
+    if (maxPrice) {
+      formData.append('maxPrice', maxPrice);
+    }
+    formData.append('scope', isPublic ? 'PUBLIC' : 'PRIVATE');
+
+    try {
+      const response = await createCard(formData);
+      if (response.status === 201) {
+        navigate(`/Card/${response.data.cardId}`);
+      }
+    } catch (error) {
+      console.log('생성 도중 오류가 발생했습니다: ', error);
+    }
   };
 
   const handleCancel = () => {
@@ -162,6 +196,19 @@ const SearchModal: React.FC<SearchModalProps> = ({ open, handleClose }) => {
       >
         <Box sx={{ mt: 3, width: '100%' }}>
           <Grid container spacing={2}>
+            <Grid item xs={12} sm={12}>
+              <TextField
+                value={newCardTitle}
+                onChange={handleNewCardTitleChange}
+                autoComplete="given-name"
+                name="newCardTitle"
+                required
+                fullWidth
+                id="newCardTitle"
+                label="카드 제목"
+                autoFocus
+              />
+            </Grid>
             <Grid item xs={12} sm={6}>
               <TextField
                 value={searchTerm}
@@ -175,11 +222,23 @@ const SearchModal: React.FC<SearchModalProps> = ({ open, handleClose }) => {
                 autoFocus
               />
             </Grid>
-            <Grid item xs={12} sm={6} />
+            <Grid item xs={12} sm={6}>
+              <FormControlLabel
+                control={
+                  <Checkbox
+                    checked={isPublic}
+                    onChange={handlePublicChange}
+                    name="isPublic"
+                  />
+                }
+                label="공개 여부"
+              />
+            </Grid>{' '}
             <Grid item xs={12} sm={6}>
               <TextField
                 value={minPrice}
                 onChange={handleMinPriceChange}
+                required
                 fullWidth
                 id="minPrice"
                 label="최소 가격"
@@ -191,6 +250,7 @@ const SearchModal: React.FC<SearchModalProps> = ({ open, handleClose }) => {
               <TextField
                 value={maxPrice}
                 onChange={handleMaxPriceChange}
+                required
                 fullWidth
                 id="maxPrice"
                 label="최대 가격"
@@ -208,7 +268,6 @@ const SearchModal: React.FC<SearchModalProps> = ({ open, handleClose }) => {
                   label="카테고리"
                   onChange={handleCategoryChange}
                 >
-                  <MenuItem value="default">----------------</MenuItem>
                   {category.map((category) => (
                     <MenuItem
                       value={category.categoryId}
@@ -243,15 +302,14 @@ const SearchModal: React.FC<SearchModalProps> = ({ open, handleClose }) => {
             </Grid>
             <Grid item xs={12} sm={4}>
               <FormControl fullWidth>
-                <InputLabel id="area-select-label">시도</InputLabel>
+                <InputLabel id="area-select-label">지역</InputLabel>
                 <Select
-                  labelId="sido-select-label"
-                  id="sido-select"
+                  labelId="area-select-label"
+                  id="area-select"
                   value={selectedSido ? selectedSido.id.toString() : ''}
-                  label="시도"
+                  label="지역"
                   onChange={handleSidoSelect}
                 >
-                  <MenuItem value="default">-------------</MenuItem>
                   {area.map((sido) => (
                     <MenuItem key={sido.id} value={sido.id}>
                       {sido.name}
@@ -302,9 +360,9 @@ const SearchModal: React.FC<SearchModalProps> = ({ open, handleClose }) => {
               type="submit"
               variant="contained"
               sx={{ mt: 3, mb: 2 }}
-              onClick={handleSearch}
+              onClick={handleCreate}
             >
-              검색
+              카드 생성
             </Button>
             <Button
               type="button"
@@ -321,4 +379,4 @@ const SearchModal: React.FC<SearchModalProps> = ({ open, handleClose }) => {
   );
 };
 
-export default SearchModal;
+export default CreateCardModal;
